+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Remove `Product-Specifications` configuration throught CMS, due to it being a breaking change. 
 
 ## [1.15.0] - 2019-04-25
 ### Changed